+++ frontend/src/store.js
@@ -56,6 +56,7 @@ const state = reactive({
         rate_limit_app_reset_time: '',
         rate_limit_app_active: '',
         allow_login_managed: '',
+        disable_alt_login: '',
         allow_character_deletion: '',
         fetch_structure_name_error_days: '',